longarr PR #2: windows-32 layout pad, copilot fixes, doc :ref: fix
Three groups of fixes for the previous round of CI + review:

1) windows-32 build failure — `static_assert: das_table.keys / .hashes offset drift`.
   Root cause: C++ `Table : Array` forces sizeof(Array)=40 on 32-bit (Array's
   uint64_t members alignment-pad it). The derived class places `keys` at
   offset 40. The flat das_table C mirror laid `keys` at offset 36 (right
   after `flags`). Mismatch only manifests on 32-bit MSVC.

   Fix: add explicit `uint32_t _pad_after_flags` to das_array and das_table
   in daScriptC.h so the flat mirror matches the C++ inheritance layout on
   every ABI. Mirror the same pad slot in the dasLLVM JIT struct (new
   JIT_ARRAY.PAD_AFTER_FLAGS / JIT_TABLE.PAD_AFTER_FLAGS enum entries +
   extra t_int32 field) so JIT'd reads/writes hit the same offsets on
   32-bit. No-op on 64-bit (natural padding already produced the same
   layout).

2) Copilot review — accepted bugs:

   - module_builtin_array.cpp builtin_array_erase (int): treated negative
     index as out-of-range but printed it as %u (uint32_t(-1) =
     4294967295). Now checks `index < 0` explicitly and prints `%d`.
   - module_builtin_array.cpp builtin_array_erase_range (int): signed
     overflow UB on `index + count` before the uint64 cast. Mirror the
     i64 fix — non-negativity check, then `uint64(index) + uint64(count)`.
   - module_builtin_runtime.cpp builtin_make_temp_array_i64: negative size
     wrapped to huge uint64 capacity. Clamp to 0 (empty view) instead.
   - daslib/builtin.das reserve(Tab, newSize:int) / (Tab, int64): negative
     newSize wrapped through `uint64(...)` to a huge value, potentially
     infinite-looping the rehash. Return early on negative input.
   - runtime_table.h TableHash::reserve: `newCapacity *= 2` doubling
     loop had no overflow guard. For sizes near 2^63 this wraps to 0 and
     infinite-loops. Throw before doubling once size exceeds 2^62.
   - aot.h vector int64 `at()` (both const variants) + das_ati_i64 +
     das_atci_i64: negative index was reported as a huge unsigned value
     (`uint64(-1) = 18446744073709551615`). Now explicit `index < 0`
     check + `%lld` signed format, matching the rest of the i64 codepath.
   - simulate_gc.cpp describe_ptr: tsize was narrowed `int(uint64)` for
     diagnostics. Widen describe_ptr's tsize parameter to uint64_t so the
     diagnostic reports the real size for >2GB containers.

3) Build doc — Sphinx undefined-label warnings on the new long_* RST docs
   I added in fc7f88b86 (I used `:ref:length`, `:ref:capacity`, etc., but
   no such labels exist in the generated docs). Drop the `:ref:` syntax,
   use plain inline-code backticks instead.

Two Copilot comments not addressed in this commit:

  - aot.h: `value[index]` vs `value[idx]` after `throw_error_ex`. Confirmed
    throw_error_ex is DAS_NORETURN_PREFIX/SUFFIX (simulate.h:500-502), so
    the fall-through path Copilot worried about is unreachable. Pure
    speculative defense-in-depth — rejected per pr_review_nitpick_filter.
  - llvm_jit_intrin.das intrinsic_builtin_length: Copilot wants a panic
    when size > INT32_MAX to match C++ builtin_array_size. The free
    intrinsic doesn't have access to the JIT visitor's get_line_info_ptr
    / get_context_param helpers required to emit the exception call;
    routing that requires the broader intrinsic-context refactor the plan
    deferred to Phase 7. TODO already in the code; documented in the
    review thread.

Local: interpreter 8926/8926, AOT 8315/8315, lint 0/19, das_array/das_table
runtime canary 13/13 (tests-cpp-small).

+++ src/simulate/simulate_gc.cpp
@@ -270,8 +270,8 @@ namespace das
             currentRange = ptrRangeStack.back();
             ptrRangeStack.pop_back();
         }
-        bool describe_ptr ( char * pa, int tsize, bool isHandle = false ) {
-            auto ssize = (tsize+15) & ~15;
+        bool describe_ptr ( char * pa, uint64_t tsize, bool isHandle = false ) {
+            auto ssize = (tsize + uint64_t(15)) & ~uint64_t(15);
             bool show = !errorsOnly;
             if ( context->stack.is_stack_ptr(pa) ) {
                 if ( show ) tp << "\tSTACK";
@@ -379,7 +379,7 @@ namespace das
             char * pa = PA->data;
             PtrRange rdata(pa, tsize);
             if ( reportHeap && tsize && markRange(rdata) ) {
-                if ( describe_ptr(pa, int(tsize)) ) {
+                if ( describe_ptr(pa, tsize) ) {
                     describeInfo(ti);
                     ReportHistory();
                     tp << "ARRAY " << getTypeInfoMangledName(ti) << "\n";
@@ -396,7 +396,7 @@ namespace das
             char * pa = PT->data;
             PtrRange rdata(pa, tsize);
             if ( reportHeap && tsize && markRange(rdata) ) {
-                if ( describe_ptr(pa, int(tsize)) ) {
+                if ( describe_ptr(pa, tsize) ) {
                     describeInfo(ti);
                     ReportHistory();
                     tp << "TABLE " << getTypeInfoMangledName(ti) << "\n";
